requirements.txt
fastapi==0.124.0
uvicorn[standard]==0.38.0
python-multipart==0.0.20
jinja2==3.1.6
aiofiles==25.1.0
sqlalchemy==2.0.44
aiosqlite==0.21.0
httpx==0.28.1
python-dotenv==1.2.1
passlib[bcrypt]==1.7.4
stripe==14.0.1
pydantic==2.12.5
pydantic-settings==2.12.0
pillow==12.0.0
